backend/requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI集成
langchain==0.0.354
openai==1.3.5
openclaw-sdk==0.1.0

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具和工具类
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# 缓存
redis==5.0.1

# 开发工具
httpx==0.25.2
tenacity==8.2.3
loguru==0.7.2